move remove brackets function to AbstractUtils
