+++ core/src/main/scala/org/dbpedia/extraction/nif/WikipediaNifExtractor.scala
@@ -4,7 +4,7 @@ import org.dbpedia.extraction.config.Config
 import org.dbpedia.extraction.config.provenance.DBpediaDatasets
 import org.dbpedia.extraction.ontology.{Ontology, OntologyProperty, RdfNamespace}
 import org.dbpedia.extraction.transform.{Quad, QuadBuilder}
-import org.dbpedia.extraction.util.{Language, RecordEntry, RecordSeverity, WikiUtil}
+import org.dbpedia.extraction.util.{AbstractUtils, Language, RecordEntry, RecordSeverity}
 import org.dbpedia.extraction.wikiparser.{Namespace, WikiPage}
 import org.dbpedia.extraction.wikiparser.impl.wikipedia.Namespaces
 import org.jsoup.nodes.{Document, Element, Node}
@@ -72,8 +72,8 @@ class WikipediaNifExtractor(
 
     if (recordAbstracts && extractionResults.section.id == "abstract" && extractionResults.getExtractedLength > 0) {
       val (cleanLongAbstract, cleanShortAbstract) = if (removeBrokenBrackets) {
-        (WikiUtil.removeBrokenBracketsInAbstracts(extractionResults.getExtractedText),
-          WikiUtil.removeBrokenBracketsInAbstracts(getShortAbstract(extractionResults)))
+        (AbstractUtils.removeBrokenBracketsInAbstracts(extractionResults.getExtractedText),
+          AbstractUtils.removeBrokenBracketsInAbstracts(getShortAbstract(extractionResults)))
       } else {
         (extractionResults.getExtractedText, getShortAbstract(extractionResults))
       }
