+++ core/src/test/scala/org/dbpedia/extraction/util/AbstractUtilsTest.scala
@@ -1,29 +1,29 @@
 package org.dbpedia.extraction.util
 import org.scalatest.FunSuite
 
-class WikiUtilsTest extends FunSuite {
+class AbstractUtilsTest extends FunSuite {
   test("test removing broken brackets function") {
     val text = "Berlin (; German: [bɛʁˈliːn] ()) is the capital and largest city of Germany by both area and population."
     val expectedText = "Berlin is the capital and largest city of Germany by both area and population."
-    val resultText = WikiUtil.removeBrokenBracketsInAbstracts(text)
+    val resultText = AbstractUtils.removeBrokenBracketsInAbstracts(text)
     assert(resultText == expectedText)
   }
   test("test removing broken brackets function with only empty brackets") {
     val text = "Berlin () is the capital and largest () city of Germany by both area and population."
     val expectedText = "Berlin is the capital and largest city of Germany by both area and population."
-    val resultText = WikiUtil.removeBrokenBracketsInAbstracts(text)
+    val resultText = AbstractUtils.removeBrokenBracketsInAbstracts(text)
     assert(resultText == expectedText)
   }
   test("test removing broken brackets function with bracket and semicolon") {
     val text = "Berlin (; German: [bɛʁˈliːn]) is the capital and largest city of Germany by both area and population."
     val expectedText = "Berlin is the capital and largest city of Germany by both area and population."
-    val resultText = WikiUtil.removeBrokenBracketsInAbstracts(text)
+    val resultText = AbstractUtils.removeBrokenBracketsInAbstracts(text)
     assert(resultText == expectedText)
   }
   test("test removing broken brackets function with bracket and comma") {
     val text = "Berlin (, German: [bɛʁˈliːn]) is the capital and largest city of Germany by both area and population."
     val expectedText = "Berlin is the capital and largest city of Germany by both area and population."
-    val resultText = WikiUtil.removeBrokenBracketsInAbstracts(text)
+    val resultText = AbstractUtils.removeBrokenBracketsInAbstracts(text)
     assert(resultText == expectedText)
   }
 }
